(PC-29934)[API] refactor: move apps flyer request to connector
