+++ api/src/pcapi/core/subscription/api.py
@@ -91,7 +91,7 @@ def activate_beneficiary_for_eligibility(
     batch.track_deposit_activated_event(user.id, deposit)
     amplitude_events.track_deposit_activation_event(user.id, deposit, fraud_check)
 
-    if "apps_flyer" in user.externalIds:  # type: ignore[operator]
+    if "apps_flyer" in user.externalIds:
         apps_flyer_job.log_user_becomes_beneficiary_event_job.delay(user.id)
 
     return user
